requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
firebase-admin==5.2.0
Flask==2.0.2
gcloud==0.18.3
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.13.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jwcrypto==1.0
lxml==4.7.1
MarkupSafe==2.0.1
msgpack==1.0.3
oauth2client==4.1.3
outcome==1.1.0
packaging==21.3
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
Pyrebase4==4.5.0
python-jwt==3.3.2
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
